+++ package.json
@@ -8,6 +8,7 @@
   },
   "dependencies": {
     "bignumber.js": "^7.2.1",
+    "n-readlines": "^1.0.0",
     "request": "^2.87.0",
     "underscore": "^1.9.1",
     "web3": "^1.0.0-beta.34"
